Some updates were made to the from hdf base converter so that it has less links to ASFF, so it can support other mappers easier.
